Refactoring to accommodate MVC pattern
