+++ .idea/libraries/Gradle__android_arch_lifecycle_runtime_1_0_0_aar.xml
@@ -1,8 +1,9 @@
 <component name="libraryTable">
   <library name="Gradle: android.arch.lifecycle:runtime:1.0.0@aar">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runtime-1.0.0.aar/aa0e446b123511abd011096582968cac/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runtime-1.0.0.aar/aa0e446b123511abd011096582968cac/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/cbd00f137ffe055a54edb27e1642d0f7/runtime-1.0.0/AndroidManifest.xml" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/cbd00f137ffe055a54edb27e1642d0f7/runtime-1.0.0/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/cbd00f137ffe055a54edb27e1642d0f7/runtime-1.0.0/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
