+++ .idea/libraries/Gradle__com_android_support_appcompat_v7_26_1_0_aar.xml
@@ -1,8 +1,9 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:appcompat-v7:26.1.0@aar">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/appcompat-v7-26.1.0.aar/1f5adddd01b7a3160cf768fb6838c834/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/appcompat-v7-26.1.0.aar/1f5adddd01b7a3160cf768fb6838c834/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/3c115bd9f6556a7430ab54a2645aa55d/appcompat-v7-26.1.0/AndroidManifest.xml" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/3c115bd9f6556a7430ab54a2645aa55d/appcompat-v7-26.1.0/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/3c115bd9f6556a7430ab54a2645aa55d/appcompat-v7-26.1.0/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
