+++ .idea/libraries/Gradle__com_android_support_constraint_constraint_layout_1_0_2_aar.xml
@@ -1,8 +1,9 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support.constraint:constraint-layout:1.0.2@aar">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/constraint-layout-1.0.2.aar/8f297e42334f3f62da28bd3d39483c1b/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/constraint-layout-1.0.2.aar/8f297e42334f3f62da28bd3d39483c1b/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/cb20b76ed225fe8217fdebe6b4cd159d/constraint-layout-1.0.2/AndroidManifest.xml" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/cb20b76ed225fe8217fdebe6b4cd159d/constraint-layout-1.0.2/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/cb20b76ed225fe8217fdebe6b4cd159d/constraint-layout-1.0.2/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
