+++ .idea/libraries/Gradle__com_android_support_design_26_1_0_aar.xml
@@ -1,8 +1,9 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:design:26.1.0@aar">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/design-26.1.0.aar/20444f3fb7d8ce504cb9c490daa69e0b/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/design-26.1.0.aar/20444f3fb7d8ce504cb9c490daa69e0b/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/20002896a77046aac10817efbb717d4b/design-26.1.0/AndroidManifest.xml" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/20002896a77046aac10817efbb717d4b/design-26.1.0/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/20002896a77046aac10817efbb717d4b/design-26.1.0/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
