+++ .idea/libraries/Gradle__com_android_support_support_compat_26_1_0_aar.xml
@@ -1,8 +1,9 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:support-compat:26.1.0@aar">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-compat-26.1.0.aar/9ce8208c5e9e96e0d929dca3e53577e5/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-compat-26.1.0.aar/9ce8208c5e9e96e0d929dca3e53577e5/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/19640d657814c0deaea3c9f8ce878766/support-compat-26.1.0/AndroidManifest.xml" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/19640d657814c0deaea3c9f8ce878766/support-compat-26.1.0/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/19640d657814c0deaea3c9f8ce878766/support-compat-26.1.0/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
