+++ .idea/libraries/Gradle__com_android_support_support_fragment_26_1_0_aar.xml
@@ -1,8 +1,9 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:support-fragment:26.1.0@aar">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-fragment-26.1.0.aar/ba309907e5bf861f1ea2332945ad8476/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-fragment-26.1.0.aar/ba309907e5bf861f1ea2332945ad8476/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/f7553f20c19b1ac99491fe60ed6b734d/support-fragment-26.1.0/AndroidManifest.xml" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/f7553f20c19b1ac99491fe60ed6b734d/support-fragment-26.1.0/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/f7553f20c19b1ac99491fe60ed6b734d/support-fragment-26.1.0/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
