+++ .idea/libraries/Gradle__com_android_support_support_media_compat_26_1_0_aar.xml
@@ -1,8 +1,9 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:support-media-compat:26.1.0@aar">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-media-compat-26.1.0.aar/d398da3bb10f97fec1a26cce7346fb84/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-media-compat-26.1.0.aar/d398da3bb10f97fec1a26cce7346fb84/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/26df0f8813a13404e327c780b3b81904/support-media-compat-26.1.0/AndroidManifest.xml" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/26df0f8813a13404e327c780b3b81904/support-media-compat-26.1.0/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/26df0f8813a13404e327c780b3b81904/support-media-compat-26.1.0/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
