+++ .idea/libraries/Gradle__com_android_support_support_v4_26_1_0_aar.xml
@@ -1,8 +1,9 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:support-v4:26.1.0@aar">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-v4-26.1.0.aar/26491ee59380eb454724efe8c8cf1b8e/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-v4-26.1.0.aar/26491ee59380eb454724efe8c8cf1b8e/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/8b2410946284cd4775e5f8b775d72d89/support-v4-26.1.0/AndroidManifest.xml" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/8b2410946284cd4775e5f8b775d72d89/support-v4-26.1.0/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/8b2410946284cd4775e5f8b775d72d89/support-v4-26.1.0/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
