+++ .idea/libraries/Gradle__com_android_support_support_vector_drawable_26_1_0_aar.xml
@@ -1,8 +1,9 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:support-vector-drawable:26.1.0@aar">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-vector-drawable-26.1.0.aar/41b2ea8b7686e6e8e979a12e84338a7d/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-vector-drawable-26.1.0.aar/41b2ea8b7686e6e8e979a12e84338a7d/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/a449755bc5e7ed79153fa9929dc7654c/support-vector-drawable-26.1.0/AndroidManifest.xml" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/a449755bc5e7ed79153fa9929dc7654c/support-vector-drawable-26.1.0/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/a449755bc5e7ed79153fa9929dc7654c/support-vector-drawable-26.1.0/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
