+++ .idea/libraries/Gradle__com_android_support_test_rules_0_5_aar.xml
@@ -1,11 +1,12 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support.test:rules:0.5@aar">
     <ANNOTATIONS>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/rules-0.5.aar/1d5445c80ecbb8e4a9e19a5a5eb9ad7e/annotations.zip!/" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/c13f3f5a68b305eb2d0eb6d45baf21dd/rules-0.5/annotations.zip!/" />
     </ANNOTATIONS>
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/rules-0.5.aar/1d5445c80ecbb8e4a9e19a5a5eb9ad7e/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/rules-0.5.aar/1d5445c80ecbb8e4a9e19a5a5eb9ad7e/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/c13f3f5a68b305eb2d0eb6d45baf21dd/rules-0.5/AndroidManifest.xml" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/c13f3f5a68b305eb2d0eb6d45baf21dd/rules-0.5/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/c13f3f5a68b305eb2d0eb6d45baf21dd/rules-0.5/res" />
     </CLASSES>
     <JAVADOC>
       <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.android.support.test/rules/0.5/71c1df91db9f063a68e2e52a8568c9f3bd3c1c73/rules-0.5-javadoc.jar!/" />
