+++ .idea/libraries/Gradle__com_android_support_transition_26_1_0_aar.xml
@@ -1,8 +1,9 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:transition:26.1.0@aar">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/transition-26.1.0.aar/dffbe562b80573e09a3c479dc841880d/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/transition-26.1.0.aar/dffbe562b80573e09a3c479dc841880d/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/e13c4a3a02d0f93e83b088e91c91216c/transition-26.1.0/AndroidManifest.xml" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/e13c4a3a02d0f93e83b088e91c91216c/transition-26.1.0/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/e13c4a3a02d0f93e83b088e91c91216c/transition-26.1.0/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
