+++ .idea/libraries/Gradle__com_mcxiaoke_volley_library_aar_1_0_0_aar.xml
@@ -1,8 +1,9 @@
 <component name="libraryTable">
   <library name="Gradle: com.mcxiaoke.volley:library-aar:1.0.0@aar">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/library-aar-1.0.0.aar/cb2e73adcdff623873c0232852491600/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/library-aar-1.0.0.aar/cb2e73adcdff623873c0232852491600/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/45eee55caa8987a3eedc953f8293d1d4/library-aar-1.0.0/AndroidManifest.xml" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/45eee55caa8987a3eedc953f8293d1d4/library-aar-1.0.0/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/45eee55caa8987a3eedc953f8293d1d4/library-aar-1.0.0/res" />
     </CLASSES>
     <JAVADOC>
       <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.mcxiaoke.volley/library-aar/1.0.0/90bda4b724359a24f920b5195aba503152e17591/library-aar-1.0.0-javadoc.jar!/" />
