+++ .idea/modules.xml
@@ -2,8 +2,8 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
-      <module fileurl="file://$PROJECT_DIR$/mylibrary.iml" filepath="$PROJECT_DIR$/mylibrary.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/MyLibrary.iml" filepath="$PROJECT_DIR$/.idea/modules/MyLibrary.iml" group="MyLibrary" />
+      <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" group="MyLibrary/app" />
     </modules>
   </component>
 </project>
