+++ app/src/main/java/com/sandhu/manny/mylibrary/AddBookActivity.java
@@ -14,6 +14,7 @@ import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.JsonObjectRequest;
 import com.android.volley.toolbox.Volley;
+import com.sandhu.manny.mylibrary.dao.DatabaseHelper;
 
 import org.json.JSONArray;
 import org.json.JSONException;
