+++ app/src/main/java/com/sandhu/manny/mylibrary/AddBookFromShelfActivity.java
@@ -1,6 +1,5 @@
 package com.sandhu.manny.mylibrary;
 
-import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
@@ -15,6 +14,7 @@ import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.JsonObjectRequest;
 import com.android.volley.toolbox.Volley;
+import com.sandhu.manny.mylibrary.dao.DatabaseHelper;
 
 import org.json.JSONArray;
 import org.json.JSONException;
