+++ app/src/main/java/com/sandhu/manny/mylibrary/ManageShelfActivity.java
@@ -16,6 +16,8 @@ import android.widget.RelativeLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.sandhu.manny.mylibrary.dao.DatabaseHelper;
+
 import java.util.ArrayList;
 
 /*
