+++ app/src/main/java/com/sandhu/manny/mylibrary/dao/DatabaseHelper.java
@@ -1,4 +1,4 @@
-package com.sandhu.manny.mylibrary;
+package com.sandhu.manny.mylibrary.dao;
 
 import android.content.ContentValues;
 import android.content.Context;
