+++ app/src/main/java/com/sandhu/manny/mylibrary/model/Volume.java
@@ -1,4 +1,4 @@
-package com.sandhu.manny.mylibrary;
+package com.sandhu.manny.mylibrary.model;
 
 public class Volume {
