+++ app/src/main/java/com/sandhu/manny/mylibrary/service/LibraryService.java
@@ -1,0 +1,4 @@
+package com.sandhu.manny.mylibrary.service;
+
+public class LibraryService {
+}
