+++ app/src/main/java/com/sandhu/manny/mylibrary/service/ShelfService.java
@@ -1,0 +1,4 @@
+package com.sandhu.manny.mylibrary.service;
+
+public class ShelfService {
+}
